refactored dataset routes
